De-duplicate PQC type helpers and unify source files.
